refactor(button): remove duplicate code
