+++ packages/aksara-ui-core/src/components/button/components/Button/components/InvisibleText.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const InvisibleText = styled('span')`
+  visibility: hidden;
+`;
+
+export default InvisibleText;
