+++ packages/aksara-ui-core/src/components/button/components/Button/components/index.ts
@@ -1,2 +1,3 @@
 export { default as IconWrapper } from './IconWrapper';
+export { default as InvisibleText } from './InvisibleText';
 export { default as LoaderCircle } from './LoaderCircle';
