+++ packages/aksara-ui-core/src/components/button/components/Button/utils/index.ts
@@ -1,2 +1,3 @@
-export { default as loadingIconSizes } from './loadingIconSizes';
+export { default as renderButtonChildren } from './renderButtonChildren';
+export { default as renderButtonIcon } from './renderButtonIcon';
 export * from './buttonUtils';
